Move HTTP Client protocol and implementation to stand alone folders representing modules
